refactor(dao/entities): update imports and implement DAO pattern
- Refactored imports to match new package structure (`dao` and `entities`).
- Implemented DAO pattern with `CategoryDAO` and `BookDAO` interfaces and their implementations.
- Enhanced `Program` to demonstrate DAO usage for saving entities.
